mail/requirements.txt
# direct dependencies
google-api-python-client==2.57.0
google-auth-oauthlib==0.5.2
pylint==2.14.5
PyYAML==6.0
Jinja2==3.1.2
# indirect dependencies
astroid==2.11.7
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
dill==0.3.5.1
google-api-core==2.8.2
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.0
platformdirs==2.5.2
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
tomlkit==0.11.4
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
